+++ python/lsst/sattle/app/constants.py
